"Update Conversations and Search components to use filteredUsers state and pass users state as prop to Search and Conversations"
This commit updates the Conversations and Search components in the Sidebar of the chat application. The Convers
