+++ client/chat-app/src/components/Sidebar/Conversations.jsx
@@ -1,25 +1,12 @@
-import { useEffect, useState } from "react";
 import Conversation from "./Conversation";
-import axios from "axios";
 import ChatSkeleton from "../../skeletons/ChatSkeleton";
 
-const Conversations = () => {
-  const [users, setUsers] = useState(null);
-  const getUsers = async () => {
-    const res = await axios.get("http://localhost:7777/api/users/getusers", {
-      withCredentials: true,
-    });
-    setTimeout(() => {
-      setUsers(res.data);
-    }, 300);
-    console.log(res.data);
-  };
-  useEffect(() => {
-    getUsers();
-  }, []);
+const Conversations = ({ users }) => {
   return (
     <>
-      {users ? (
+      {users === "No User Found" ? (
+        <h1>No user Found</h1>
+      ) : users ? (
         users.map((user) => <Conversation key={user._id} user={user} />)
       ) : (
         <>
